backend/requirements/base.txt
django==5.1
psycopg[binary]==3.2.1
python-decouple==3.7
django-allauth[socialaccount]
djangorestframework==3.15.2
djangorestframework-simplejwt==5.3.1
django-cors-headers==4.4.0
cryptography==43.0.0
dj-rest-auth==6.0.0
pillow==10.4.0
python-magic==0.4.27
querystring-parser==1.2.4
pytz==2024.1

channels==4.1.0
channels-redis==4.2.0
gunicorn==23.0.0
requests==2.32.3
google-auth==2.23.0

django-cleanup==8.1.0
django-simple-history==3.4.0
openpyxl==3.1.2

celery==5.4.0
flower==2.0.1
celery[redis]

phonenumbers==8.13.29
django-phonenumber-field==7.3.0
django-countries==7.6.1